I fixed the survey and survey details pages disappearing by refactoring SurveyForm.
